requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
mysql-connector-python
python-dotenv  
